Rename CodeWrapper to CodeBlock
