+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './codeWrapper/CodeWrapper';
+export * from './codeBlock/CodeBlock';
 export * from './pageHeader/PageHeader';
 export * from './groupedAccordions/GroupedAccordions';
 export * from './header/Header';
